+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.2.8.dev (development stage/unreleased)
+### Added
+- get_monitoring_status_icinga() in manager class
 
 ## 1.2.8
 ### Added